// packages/components/src/theme/types.ts
export type Intent =
  | 'primary'
  | 'secondary'
  | 'light'
  | 'dark'
  | 'info'
  | 'success'
  | 'warning'
  | 'danger'

/** UI Element size, typically ranging from xs to xl. */
export type ElementSize = 'xs' | 'sm' | 'md' | 'lg' | 'xl'

export type Alignment = 'center' | 'left' | 'right' | 'middle' | 'start' | 'end'